+++ src/impl/index.ts
@@ -1,2 +1,3 @@
 export * from './files-to-test-string';
 export * from './read-fake-files';
+export * from './find-fs-test-case-directories';